chore: remove obsolete combined files
