+++ Models/FormFieldConfigDto.cs
@@ -36,9 +36,3 @@ public class FormFieldConfigDto
     
     public string DATA_TYPE { get; set; }
 }
-
-public sealed record FieldConfigData(
-    List<FormFieldConfigDto> FieldConfigs,
-    List<FormFieldValidationRuleDto> ValidationRules,
-    List<FORM_FIELD_DROPDOWN> DropdownConfigs,
-    List<FORM_FIELD_DROPDOWN_OPTIONS> DropdownOptions);
