+++ Models/Option.cs
@@ -1,0 +1,8 @@
+namespace DynamicForm.Models;
+
+public class Option
+{
+    public string label { get; set; }
+    public string value { get; set; }
+}
+
