+++ ViewModels/DbColumnInfo.cs
