+++ ViewModels/DropdownAnswerDto.cs
@@ -1,0 +1,9 @@
+namespace DynamicForm.Models;
+
+public class DropdownAnswerDto
+{
+    public string RowId { get; set; } = default!;
+    public Guid FieldId { get; set; }
+    public Guid OptionId { get; set; }
+}
+
