+++ ViewModels/FormDataCell.cs
@@ -1,0 +1,8 @@
+namespace DynamicForm.Models;
+
+public class FormDataCell
+{
+    public string ColumnName { get; set; } = string.Empty;
+    public object? Value { get; set; }
+}
+
