+++ ViewModels/FormSubmissionInputModel.cs
@@ -13,15 +13,3 @@ public class FormSubmissionInputModel
     public string? TargetTableToUpsert { get; set; }
     public List<FormInputField> InputFields { get; set; } = new();
 }
-
-/// <summary>
-/// Represent single field input for submission.
-/// </summary>
-public class FormInputField
-{
-    public Guid FieldConfigId { get; set; }
-    public string? Value { get; set; }
-    public string? Column { get; set; }
-
-    public string? SOURCE_TABLE { get; set; }
-}
